Dry up byte[] backed AbstractBigArray implementations (#108636)
Dry up the code for all the `byte[]` backed versions of the big array
a little. The motivation here (outside of making the code drier now) is
to follow-up with possible optimizations for sparsely populated variants
of this thing.
