+++ server/src/main/java/org/elasticsearch/common/util/AbstractBigArray.java
@@ -19,7 +19,7 @@ import java.util.Arrays;
 /** Common implementation for array lists that slice data into fixed-size blocks. */
 abstract class AbstractBigArray extends AbstractArray {
 
-    private final PageCacheRecycler recycler;
+    protected final PageCacheRecycler recycler;
     private Recycler.V<?>[] cache;
 
     private final int pageShift;
@@ -93,30 +93,12 @@ abstract class AbstractBigArray extends AbstractArray {
         return array;
     }
 
-    private <T> T registerNewPage(Recycler.V<T> v, int page, int expectedSize) {
+    protected <T> T registerNewPage(Recycler.V<T> v, int page, int expectedSize) {
         cache = grow(cache, page + 1);
         assert cache[page] == null;
         cache[page] = v;
         assert Array.getLength(v.v()) == expectedSize;
         return v.v();
-    }
-
-    protected final byte[] newBytePage(int page) {
-        if (recycler != null) {
-            final Recycler.V<byte[]> v = recycler.bytePage(clearOnResize);
-            return registerNewPage(v, page, PageCacheRecycler.BYTE_PAGE_SIZE);
-        } else {
-            return new byte[PageCacheRecycler.BYTE_PAGE_SIZE];
-        }
-    }
-
-    protected final Object[] newObjectPage(int page) {
-        if (recycler != null) {
-            final Recycler.V<Object[]> v = recycler.objectPage();
-            return registerNewPage(v, page, PageCacheRecycler.OBJECT_PAGE_SIZE);
-        } else {
-            return new Object[PageCacheRecycler.OBJECT_PAGE_SIZE];
-        }
     }
 
     protected final void releasePage(int page) {
@@ -131,40 +113,6 @@ abstract class AbstractBigArray extends AbstractArray {
         if (recycler != null) {
             Releasables.close(cache);
             cache = null;
-        }
-    }
-
-    /**
-     * Fills an array with a value by copying it to itself, increasing copy ranges in each iteration
-     */
-    protected static final void fillBySelfCopy(byte[] page, int fromBytes, int toBytes, int initialCopyBytes) {
-        for (int pos = fromBytes + initialCopyBytes; pos < toBytes;) {
-            int sourceBytesLength = pos - fromBytes; // source bytes available to be copied
-            int copyBytesLength = Math.min(sourceBytesLength, toBytes - pos); // number of bytes to actually copy
-            System.arraycopy(page, fromBytes, page, pos, copyBytesLength);
-            pos += copyBytesLength;
-        }
-    }
-
-    /**
-     * Bulk copies array to paged array
-     */
-    public void set(long index, byte[] buf, int offset, int len, byte[][] pages, int shift) {
-        assert index + len <= size();
-        int pageIndex = pageIndex(index);
-        final int indexInPage = indexInPage(index);
-        if (indexInPage + len <= pageSize()) {
-            System.arraycopy(buf, offset << shift, pages[pageIndex], indexInPage << shift, len << shift);
-        } else {
-            int copyLen = pageSize() - indexInPage;
-            System.arraycopy(buf, offset << shift, pages[pageIndex], indexInPage, copyLen << shift);
-            do {
-                ++pageIndex;
-                offset += copyLen;
-                len -= copyLen;
-                copyLen = Math.min(len, pageSize());
-                System.arraycopy(buf, offset << shift, pages[pageIndex], 0, copyLen << shift);
-            } while (len > copyLen);
         }
     }
