+++ server/src/main/java/org/elasticsearch/common/util/BigByteArray.java
@@ -26,20 +26,13 @@ import static org.elasticsearch.common.util.PageCacheRecycler.PAGE_SIZE_IN_BYTES
  * Byte array abstraction able to support more than 2B values. This implementation slices data into fixed-sized blocks of
  * configurable length.
  */
-final class BigByteArray extends AbstractBigArray implements ByteArray {
+final class BigByteArray extends AbstractBigByteArray implements ByteArray {
 
     private static final BigByteArray ESTIMATOR = new BigByteArray(0, BigArrays.NON_RECYCLING_INSTANCE, false);
 
-    private byte[][] pages;
-
     /** Constructor. */
     BigByteArray(long size, BigArrays bigArrays, boolean clearOnResize) {
-        super(BYTE_PAGE_SIZE, bigArrays, clearOnResize);
-        this.size = size;
-        pages = new byte[numPages(size)][];
-        for (int i = 0; i < pages.length; ++i) {
-            pages[i] = newBytePage(i);
-        }
+        super(BYTE_PAGE_SIZE, bigArrays, clearOnResize, size);
     }
 
     @Override
